vesa: Make __vesacon_i915resolution() a stub under EFI
We can't use the code in __vesacon_i915resolution() under EFI, it
leads to hangs. Provide a stub implementation in efi/vesa.c that just
returns an error.

We don't use the usual idiom of moving the function into the
'firmware' structure because that would require the i915 code to move
into the core/.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>