efi: Fix screen clearing and cursor position
The logic to figure out when to clear the screen was broken, leading
to the screen sometimes containing garbage. x0,y0 are the coordinates
of the upper left part of the screen, and x1,y1 the lower right. x1,y1
will be the position of our cursor, which we can ignore, but we must
use x0,y0 to figure out when to clear the screen and not simply set
the cursor position.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>