efi: Greatly simplify the EFI Scan Code handling
By adding the EFI Scan Codes to 'keycodes' in get_key.c, prefixed with
the escape sequence '\0' to avoid clashes with ASCII characters, we
can have all the input key decoding done in one place, allowing us to
delete efi/keymap.h.

Luckily there are no conflicts with the combined BIOS codes already
present in 'keycodes', though if we were to add the EFI scan code for
the pause key (0x48) that would conflict with the existing BIOS code
for KEY_UP. Still, it shouldn't be necessary to add any more scan
codes because it doesn't make sense for things like the mute key to be
handled by Syslinux.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>