pxe: Pass absolute path to pxe_chdir()
We may have moved out of the root directory when calling get_prefix()
and so need a way to create an absolute path.

Historically, in versions prior to 5.xx we would call
pxe_open_config() while we were still in the root directory ("::"),
and so, we could construct a relative path to the config file.

However, because we need to load ldlinux.c32 before opening the config
file, which may involve us changing directory, we need a way of
constructing an absolute path to the config file in get_prefix().

Signed-off-by: Matt Fleming <matt.fleming@intel.com>