ldlinux: use platform filename extension
Users are starting to package all Syslinux versions on one boot medium
and are running into problems because all versions of ldlinux have the
same filename, ldlinux.c32. The version of ldlinux.c32 compiled for bios
will not work with efi32 or efi64, and vice versa. What we need is a way
for the Syslinux core to know exactly which filename to load.

Use the following extensions for the corresponding platforms,

  .c32 - bios
  .e32 - efi32
  .e64 - efi64

Now all ldlinux variants can be kept in the same directory without
causing any conflicts.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>