efi: shrink the size of syslinux.efi
syslinux.efi currently includes the .bss section of the ELF shared
object, syslinux.so, as zero'd space in the file. This dramatically
increases the size of syslinux.efi.

Use the 'initialized_data_sz' field in the MSDOS/COFF header to
differentiate the on-disk and in-memory file size, and to make the
firmware loader zero the .bss for us.

efi/wrapper now only copies upto the offset of the .bss section from
syslinux.so when building syslinux.efi.

With this change syslinux.efi shrinks from 2.4M to 168K.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>