serial: fix SERIAL directive for ports > 0
There is a bug in the SERIAL directive parsing code that means that
using a port other than the first (SERIAL 0) results in a non-working
serial console. We need to use the serial port number from the config
file to offset the BIOS I/O port base address, otherwise we always
lookup the I/O port address for the first serial port instead of the one
the user specified.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>