font: write to 'fontbuf', not random memory
commit be5a345d385d ("font: load data as a single block") cleaned up the
loop for reading into the font buffer but made the mistake of dropping
the assignment to 'p', which means that the _fread() call is writing
through a garbage pointer. Since 'p' was only used to keep track while
writing into the font buffer in a loop, we can now delete it and
reference 'fontbuf' directly.

Also delete the unused variable 'i'.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>