cli: don't disable linewrap when redrawing the cmdline
It messes things up when we want to use printf(), for example when
printing labels after the user has hit the TAB key.

This fixes a bug where labels would be printed on a single line on the
screen (with some being displayed off screen) when they should have been
printed across multiple lines.

I can't think of a reason to disable linerap when redrawing the command
line - so let's not.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>