efi: implement Linux kernel handover protocol support
The handover protocol is the preferred method of booting kernels on EFI
because it allows workarounds for various firmware bugs to be contained
in one place and applied irrespective of the chosen bootloader. Use it
if available, but ensure that we fallback to the legacy boot method.

Also, update the linux_header structure with recent changes made in the
kernel source.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>