efi, udp: use single local port for each connection
The TFTP protocol uses the local port as an idenitifer during a transfer
(TID), which means that once we've established a TFTP connection, we
must ensure we reuse the same local port number in each packet. Failure
to do so is an error, which causes the TFTP server to send an error
packet.

From RFC 1350 - THE TFTP PROTOCOL (REVISION 2), Section 4,

 In the next step, and in all succeeding steps, the hosts should make
 sure that the source TID matches the value that was agreed on in
 steps 1 and 2.  If a source TID does not match, the packet should be
 discarded as erroneously sent from somewhere else.  An error packet
 should be sent to the source of the incorrect packet, while not
 disturbing the transfer.

Once the UDPv4 protocol driver has been assigned a local port number
(which happens on the first core_udp_connect()) reuse that number until
core_udp_close() time.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>