bios, font: zero input registers to get font
Users are reporting font issues under VirtualBox. Try this as a
speculative fix since the old 5.x behaviour was to pass in a partially
zero'd register set. Incidentally, the input regs are also used for
output, so at least we now avoid potentially constructing a pointer to
the font buffer using garbage register values.

Cc: H. Peter Anvin <hpa@zytor.com>
Signed-off-by: Matt Fleming <matt.fleming@intel.com>