efi: only read ADV when booting from disk
We aren't expecting to be able to access ldlinux.sys when booting over
PXE. Follow the pre-5.00 behaviour and simply initialise the ADV buffer
without even attempting to open the file. Also, delete all references to
EXTLINUX, we do not support it under EFI.

Without this patch we hang in efi_open() when booting PXELINUX because
no root volume is set as we're not booting from disk. This regression
was caused by commit a17fdfb8 ("com32: Catch up with GCC changes to
ctor/dtor funcs") because now constructors are actually being invoked on
EFI and efi_adv_init() is called by __syslinux_init().

Signed-off-by: Matt Fleming <matt.fleming@intel.com>