tests: make mmap_entry globally available to unit tests
There's a bunch of places that need to deal with mmap_entry entries when
doing tests, so make the structure readily available. Since we're going
to get a collection of header files that could be considered the
"unit test infrastructure" we might as well create a new directory for
them.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>