tests: delete test executables after each run
The unit tests are designed to be small and should build very quickly,
so there should be no downside to rebuilding them for every run. The
upside is that we don't litter our build trees with target executables
and we don't need to explicitly list dependencies in Makefiles to ensure
our tests get rebuilt whenever a dependency changes.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>