tests: unit test bios_boot_linux()
The constraints for allocating the kernel cmdline buffer under bios are
pretty involved and filled with historic rules. Unit test the bios linux
loader to ensure we never violate any of them, while at the same time
making sure we actually find a usable chunk of memory.

This commit is designed to test the changes in commit 77cadda8
("load_linux: dynamically calculate the cmdline region").

Signed-off-by: Matt Fleming <matt.fleming@intel.com>