cmenu: the order of entries in $(LIBMENU) is vital
Unresolved symbols in an object file cannot be resolved from shared
libraries that are listed earlier in $(LIBMENU). We need to put the
shared libraries at the end of $(LIBMENU).

commit 1408e6ca ("Add per-firmware object directory support") appears to
have broken this rule, which resulted in hdt.c32 (among other modules)
failing to load with,

     "Undef symbol FAIL: symbol console_ansi_raw"

Add a simple regression test that loads hdt.c32.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>