pxe, ftp: Don't convert an IP address to host byte order
FTP passes an IP address and port number as decimal string encoded
comma-separated octets.  The first four are the IP number and the
second two the port.  We consider port numbers as ordinary integers
and they should be in host byte order when passed into
core_tcp_connect(), but IP addresses are simply kept in network byte
order at all times and thus should not be converted.

Signed-off-by: H. Peter Anvin <hpa@linux.intel.com>