efi: Make gnu-efi build scripts more robust
Modify the scripts to check that they're called with the correct number
of arguments and error out otherwise after printing some helpful info.

This change also stops relying on passing arguments through environment
variables and instead passes them explicitly to the scripts, which is
definitely more robust, and handles the case where the scripts are
invoked directly.

Of course, now that the scripts can be invoked directly we need to
regard any input as hostile and exit immediately on error.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>