acpi: Removing memset of buffer
There is no need to memset the buffer since the strncat will end the
string with a \0. The memset was also almost wrong as doing a sizeof()
on a char * could return 1 if buff was malloc'ed.

We had chance as all the current calls are done with static buffers.

Removing this memset call will make things clearer but also will prevent
compilation warnings like :

com32/gpllib/acpi/acpi.c:38:29: warning: argument to ‘sizeof’ in ‘memset’ call is the same expression as the
destination; did you mean to provide an explicit length? [-Wsizeof-pointer-memaccess]
     memset(buffer, 0, sizeof(buffer));