efi: Tag __syslinux_adv* with __export
commit 415d571 ("adv: Remove double defintion") introduced a regression
for the EFI boot loader. The commit is correct, but should have tagged
the __syslinux_adv_ptr and __syslinux_adv_size data objects in the EFI
core as __export.

This change is required because symbols in the EFI code are not global
by default, and so, when ldlinux.{e32,e64} tries to bind to the adv
symbols it fails.

The following demonstrates the ELF symbol changes,

Before:
  659: 0000000000034f40     8 OBJECT  LOCAL  DEFAULT   16 __syslinux_adv_ptr

After:
  957: 0000000000036fe0     8 OBJECT  GLOBAL DEFAULT   16 __syslinux_adv_ptr

Signed-off-by: Matt Fleming <matt.fleming@intel.com>