build: sort sources to build in a more deterministic way
It has been observed that binaries contents
are depending on the order of linked objects.

This order is caused by GNU make's wildcard function
and the position of sources on filesystem.

This change tries to prevent this kind of randomness.

Also consider building using -j1 flag
to make it even more reproductible.

Change-Id: Ie8eee7f336e6f1fa2863c4150d967afd15519f1d
Bug: http://bugzilla.syslinux.org/show_bug.cgi?id=57#related
Signed-off-by: Philippe Coval <philippe.coval at open.eurogiciel.org>