hdt: fix sizeof(char *) misuse
The code was passing sizeof(char *) - not the length of the buffer - to
memset.  Change the function to take the length of the buffer as a
parameter.

Fixes the warning: argument to 'sizeof' in 'memset' call is the same
expression as the destination; did you mean to provide an explicit
length?

Signed-off-by: Jonathan Boeing <jonathan.n.boeing@gmail.com>