hdt: fix sizeof(char *) misuse
The code was passing sizeof(const char *) - not the length of the string
or the size of the buffer - as the length to strncmp().  These two cases
don't need the length, so switch to strcmp.

Fixes the warning: argument to 'sizeof' in 'strncmp' call is the same
expression as the second source; did you mean to provide an explicit
length?

Signed-off-by: Jonathan Boeing <jonathan.n.boeing@gmail.com>