hdt: Avoid false-positive single command detection
Since commit 1697594b61f9a8f9d092996afc0e2c80bbb2a20a, some commands are
said to be "nomodule" like "say".

This patch was adding a check if the nomodule flag was set but didn't
checked that the associated structure did exist leading to false
positive detection.

As a result, the commands were not executed meaning the CLI was unsuable
since ... 3 years.... *shame*

This commit simply avoid considering the nomodule flag if the structure
is not allocated